Cambio en mi componente de Product
