+++ src/app/grijalba/components/Add.jsx
@@ -1,13 +1,13 @@
 "use client";
 
-function btnCarrito() {
+function addSuccess() {
   alert("Producto agregado al carrito");
 }
 
 function AddCarrito() {
   return (
-    <button className="btn btn-success" onClick={btnCarrito}>
-      🛒
+    <button className="btn btn-success" onClick={addSuccess}>
+      + 🛒
     </button>
   );
 }
